# solutions/issuedesk-web/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "issuedesk API — API-key endpoints",
    "version": "0.1.0",
    "description": "The subset of the issuedesk API that an **API key** can call.\n\nAPI keys are read-only — they may call any `GET` endpoint — **with one family of exceptions**: managing comments. A key may post (`POST …/comments`), edit (`PATCH …/comments/{id}`) and delete (`DELETE …/comments/{id}`) comments. Everything else (sign-in, creating/editing/deleting issues, user and key management) is reserved for the browser app and is not shown here.\n\n## Authenticating\n\nSend your key as `X-API-Key: idk_…` (or `Authorization: Bearer idk_…`). API keys are exempt from the browser-lock, so curl/agents need no extra headers. A key acts on behalf of the user who created it, so any comment it posts is attributed to that user; it may edit or delete its own comments (and, as it carries admin rights, any other comment).\n\nThe same endpoints also work from the signed-in browser app via a JWT bearer token."
  },
  "servers": [{ "url": "/", "description": "This host" }],
...
  "components": {
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "An issuedesk API key (idk_…)."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "A JWT from the signed-in browser app (also accepted on these endpoints)."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "not_found" },
          "message": { "type": "string", "example": "issue not found" }
        }
      },
...
      "Attachment": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "issueId": { "type": "string", "format": "uuid" },
          "filename": { "type": "string" },
          "storedPath": { "type": "string" },
          "sizeBytes": { "type": "integer", "format": "int64" },
          "mimeType": { "type": "string" },
          "uploadedBy": { "type": "string", "format": "uuid" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
  "security": [{ "apiKeyAuth": [] }, { "bearerAuth": [] }],
  "paths": {
...
    "/api/tickets/{slug}/attachments": {
      "get": {
        "tags": ["tickets"],
        "summary": "List a ticket's attachments by public key",
        "description": "Slug-addressed counterpart to GET /api/issues/{id}/attachments. Each item's `id` downloads via GET /api/attachments/{id}.",
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" }, "example": "WAT-1" }
        ],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Attachment" } } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "post": {
        "tags": ["tickets"],
        "summary": "Upload an attachment to a ticket by public key",
        "description": "Slug-addressed counterpart to POST /api/issues/{id}/attachments, so a relay client (API key) can attach a file when commenting back with only the key from the ticket URL. Send multipart/form-data with a single file field. The returned `id` downloads via GET /api/attachments/{id}.",
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" }, "example": "WAT-1" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": { "file": { "type": "string", "format": "binary" } }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Created attachment", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Attachment" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "413": { "description": "File exceeds the upload size limit", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  }
}